fix : 밖으로 clickedCity 상태 나오도록
